Include LoadBalancer and EC2 classes to split methods and make easier to deal with dependencies on AWS Resources
