+++ lib/python/remove_route53_stateless_resources.py
@@ -1,7 +1,10 @@
-# Script used to exclude stateless route53 records during Terraform Destroy, records created by external-dns
+"""
+Script used to exclude stateless route53 records during
+Terraform Destroy, records created by external-dns
+"""
 
+import sys
 import boto3
-import json
 
 client = boto3.client('route53')
 raw_hosted_zones = client.list_hosted_zones()
@@ -9,12 +12,15 @@ external_dns_managed_record = []
 stateless_dns_record = []
 
 def enumerate_records():
+    """Enumerate Route53 records based on hosted zones"""
     for zone in raw_hosted_zones["HostedZones"]:
         zone_id = zone["Id"]
         zone_records = client.list_resource_record_sets(HostedZoneId=str(zone_id))
         for record in zone_records["ResourceRecordSets"]:
             try:
-                if ("external-dns" in str(record["ResourceRecords"][0]["Value"]) and str(record["Type"]) == "TXT" and "cname" not in str(record["Name"])):
+                if ("external-dns" in str(record["ResourceRecords"][0]["Value"])
+                        and str(record["Type"]) == "TXT"
+                        and "cname" not in str(record["Name"])):
                     record["ZoneId"] = zone_id
                     external_dns_managed_record.append(record)
             except Exception:
@@ -27,6 +33,7 @@ def enumerate_records():
                 stateless_dns_record.append(record)
 
 def delete_record(json_record):
+    """Delete Route53 record"""
     zone_id = json_record["ZoneId"]
     json_payload = {
         'Changes': [
@@ -60,4 +67,4 @@ if __name__ == "__main__":
             delete_record(record)
     except Exception as ex:
         print(ex)
-        exit(0)
+        sys.exit(0)
